backend/requirements.txt
# ==================== FRAMEWORK WEB ====================
Flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.3
gunicorn==23.0.0

# ==================== PROCESAMIENTO DE DATOS ====================
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
xlsxwriter==3.2.0

# ==================== PDF PROCESSING ====================
# Extracción de texto y tablas
pdfplumber==0.11.4
PyMuPDF==1.24.13
pdf2image==1.17.0

# Camelot - instalar sin dependencias CV problemáticas
camelot-py==0.11.0

# ==================== PROCESAMIENTO DE IMÁGENES ====================
Pillow==11.0.0
pytesseract==0.3.13
opencv-python==4.10.0.84

# ==================== UTILIDADES ====================
python-dotenv==1.0.1

# ==================== DEPENDENCIAS DE CAMELOT (manual) ====================
pypdf==4.3.1
tabulate==0.9.0
ghostscript==0.7

# ==================== NOTAS ====================
# IMPORTANTE PARA WINDOWS:
# 1. Instalar Ghostscript desde: https://www.ghostscript.com/download/gsdnld.html
#    Descargar: "Ghostscript 10.x for Windows (64 bit)"
#    Agregar al PATH: C:\Program Files\gs\gs10.xx\bin
#
# 2. Para Tesseract OCR (opcional):
#    Descargar desde: https://github.com/UB-Mannheim/tesseract/wiki
#    Instalar y agregar al PATH: C:\Program Files\Tesseract-OCR
#
# 3. Para instalar:
#    pip install -r requirements.txt